Fix circular deps
